requirements.txt
numpy==2.4.1
onnxruntime==1.23.2
onnxruntime_gpu==1.23.2
opencv_contrib_python==4.12.0.88
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
Pillow==12.1.0
torchvision==0.21.0+cu124
